PERFORMANCE BOOOST!!!!1!11!
For drawing 1000 blocks, got from 67ms down to 8 ms:
1. convert pict to bitmap
2. memoize bitmap (using memo library)

Drawing a bitmap is faster than drawing rectangles each time,
and by using memoization, we only need to convert to bitmap once.
Then we use the bitmap from that point on.
Plus, because we use memoization, we don't need to precompute
anything, the bitmaps will be computed once as new blocks appear.